separate out GlobalContext into smaller static variable
