+++ crates/fluvio-spu/src/replication/follower/state.rs
@@ -13,8 +13,9 @@ use dataplane::record::RecordSet;
 use dataplane::Offset;
 use fluvio_storage::{FileReplica, StorageError, ReplicaStorage, ReplicaStorageConfig};
 use fluvio_types::SpuId;
+use crate::replication::FileReplicaContext;
 use crate::replication::leader::ReplicaOffsetRequest;
-use crate::core::{FileGlobalContext, config};
+use crate::core::{config};
 use crate::storage::SharableReplicaStorage;
 
 use super::controller::FollowerGroups;
@@ -74,7 +75,7 @@ impl FollowersState<FileReplica> {
     /// otherwise check if there is existing state, if exists return none otherwise create new
     pub async fn add_replica(
         self: Arc<Self>,
-        ctx: &FileGlobalContext,
+        ctx: &FileReplicaContext,
         replica: Replica,
     ) -> Result<Option<FollowerReplicaState<FileReplica>>, StorageError> {
         let leader = replica.leader;
