+++ crates/fluvio-spu/src/replication/leader/replica_state.rs
@@ -690,7 +690,8 @@ mod test_leader {
     use dataplane::batch::BatchRecords;
     use dataplane::fixture::{create_recordset};
 
-    use crate::core::{spu_local_store, follower_notifier, sync_follower_update, config};
+    use crate::core::{spu_local_store, config};
+    use crate::replication::{ReplicaContext, sync_follower_update, follower_notifier};
     use crate::{
         config::{SpuConfig},
     };
@@ -888,7 +889,6 @@ mod test_leader {
 
     #[fluvio_future::test]
     async fn test_update_leader_from_followers() {
-        use crate::core::{GlobalContext};
         use fluvio_controlplane_metadata::spu::{SpuSpec};
 
         let leader_config = SpuConfig {
@@ -900,8 +900,7 @@ mod test_leader {
             SpuSpec::new_private_addr(5001, 9001, "localhost".to_owned()),
             SpuSpec::new_private_addr(5002, 9002, "localhost".to_owned()),
         ];
-        let _gctx: Arc<GlobalContext<MockStorage>> =
-            GlobalContext::new_shared_context(leader_config);
+        let _gctx: Arc<ReplicaContext<MockStorage>> = ReplicaContext::new_shared_context();
         spu_local_store().sync_all(specs);
         sync_follower_update().await;
