+++ crates/fluvio-spu/src/services/internal/mod.rs
@@ -7,7 +7,7 @@ use tracing::info;
 use fluvio_service::FluvioApiServer;
 use service_impl::InternalService;
 
-use crate::core::DefaultSharedGlobalContext;
+
 use crate::core::local_spu_id;
 
 pub use self::fetch_stream_request::FetchStreamRequest;
@@ -15,16 +15,19 @@ pub use self::fetch_stream_request::FetchStreamResponse;
 pub use self::api::SPUPeerApiEnum;
 pub use self::api::SpuPeerRequest;
 
+#[derive(Clone,Debug)]
+pub(crate) struct InternalServiceImpl;
+
 pub(crate) type InternalApiServer =
-    FluvioApiServer<SpuPeerRequest, SPUPeerApiEnum, DefaultSharedGlobalContext, InternalService>;
+    FluvioApiServer<SpuPeerRequest, SPUPeerApiEnum, InternalServiceImpl, InternalService>;
 
 // start server
-pub fn create_internal_server(addr: String, ctx: DefaultSharedGlobalContext) -> InternalApiServer {
+pub fn create_internal_server(addr: String) -> InternalApiServer {
     info!(
         "starting SPU: {} at internal service at: {}",
         local_spu_id(),
         addr
     );
 
-    FluvioApiServer::new(addr, ctx, InternalService::new())
+    FluvioApiServer::new(addr, InternalServiceImpl{}, InternalService::new())
 }
