+++ crates/fluvio-spu/src/services/public/tests/stream_fetch.rs
@@ -34,10 +34,11 @@ use fluvio_spu_schema::server::stream_fetch::LegacySmartModulePayload;
 use fluvio_spu_schema::server::stream_fetch::{DefaultStreamFetchRequest};
 use crate::{
     core::{
-        GlobalContext, spu_local_store, replica_localstore, smartmodule_localstore,
-        follower_notifier, status_update_owned, config,
+        spu_local_store, replica_localstore, smartmodule_localstore, status_update_owned, config,
+        initialize,
     },
     services::public::tests::create_filter_records,
+    replication::{ReplicaContext, default_replica_ctx, follower_notifier},
 };
 use crate::config::SpuConfig;
 use crate::replication::leader::LeaderReplicaState;
@@ -103,15 +104,17 @@ async fn test_stream_fetch_basic() {
     let addr = format!("127.0.0.1:{}", port);
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path;
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
 
     let client_socket =
         MultiplexerSocket::new(FluvioSocket::connect(&addr).await.expect("connect"));
+
+    let ctx = default_replica_ctx();
 
     // perform for two versions
     for version in 10..11 {
@@ -300,18 +303,13 @@ async fn legacy_test<Fut, TestFn>(
     test_fn: TestFn,
 ) where
     Fut: Future<Output = ()>,
-    TestFn: FnOnce(
-        Arc<GlobalContext<FileReplica>>,
-        PathBuf,
-        Option<LegacySmartModulePayload>,
-        Option<SmartModuleInvocation>,
-    ) -> Fut,
+    TestFn: FnOnce(PathBuf, Option<LegacySmartModulePayload>, Option<SmartModuleInvocation>) -> Fut,
 {
     let test_path = temp_dir().join(test_name);
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
+    initialize(spu_config);
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
     let wasm = read_wasm_module(module_name);
     let wasm_payload = LegacySmartModulePayload {
         wasm: SmartModuleWasmCompressed::Raw(wasm),
@@ -319,7 +317,7 @@ async fn legacy_test<Fut, TestFn>(
         ..Default::default()
     };
 
-    test_fn(ctx, test_path, Some(wasm_payload), None).await
+    test_fn(test_path, Some(wasm_payload), None).await
 }
 
 async fn adhoc_test<Fut, TestFn>(
@@ -329,18 +327,13 @@ async fn adhoc_test<Fut, TestFn>(
     test_fn: TestFn,
 ) where
     Fut: Future<Output = ()>,
-    TestFn: FnOnce(
-        Arc<GlobalContext<FileReplica>>,
-        PathBuf,
-        Option<LegacySmartModulePayload>,
-        Option<SmartModuleInvocation>,
-    ) -> Fut,
+    TestFn: FnOnce(PathBuf, Option<LegacySmartModulePayload>, Option<SmartModuleInvocation>) -> Fut,
 {
     let test_path = temp_dir().join(test_name);
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
     let wasm = zip(read_wasm_module(module_name));
     let smartmodule = SmartModuleInvocation {
         wasm: SmartModuleInvocationWasm::AdHoc(wasm),
@@ -348,7 +341,7 @@ async fn adhoc_test<Fut, TestFn>(
         ..Default::default()
     };
 
-    test_fn(ctx, test_path, None, Some(smartmodule)).await
+    test_fn(test_path, None, Some(smartmodule)).await
 }
 
 async fn predefined_test<Fut, TestFn>(
@@ -358,18 +351,13 @@ async fn predefined_test<Fut, TestFn>(
     test_fn: TestFn,
 ) where
     Fut: Future<Output = ()>,
-    TestFn: FnOnce(
-        Arc<GlobalContext<FileReplica>>,
-        PathBuf,
-        Option<LegacySmartModulePayload>,
-        Option<SmartModuleInvocation>,
-    ) -> Fut,
+    TestFn: FnOnce(PathBuf, Option<LegacySmartModulePayload>, Option<SmartModuleInvocation>) -> Fut,
 {
     let test_path = temp_dir().join(test_name);
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
     load_wasm_module(module_name);
     let smartmodule = SmartModuleInvocation {
         wasm: SmartModuleInvocationWasm::Predefined(module_name.to_owned()),
@@ -377,7 +365,7 @@ async fn predefined_test<Fut, TestFn>(
         ..Default::default()
     };
 
-    test_fn(ctx, test_path, None, Some(smartmodule)).await
+    test_fn(test_path, None, Some(smartmodule)).await
 }
 
 const FLUVIO_WASM_FILTER: &str = "fluvio_wasm_filter";
@@ -427,7 +415,6 @@ async fn test_stream_fetch_filter_generic() {
 }
 
 async fn test_stream_fetch_filter(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -437,7 +424,7 @@ async fn test_stream_fetch_filter(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -454,7 +441,10 @@ async fn test_stream_fetch_filter(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -629,7 +619,6 @@ async fn test_stream_fetch_filter_individual_generic() {
 }
 
 async fn test_stream_fetch_filter_individual(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -639,7 +628,7 @@ async fn test_stream_fetch_filter_individual(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -653,7 +642,10 @@ async fn test_stream_fetch_filter_individual(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -763,7 +755,6 @@ async fn test_stream_filter_error_fetch_generic() {
 }
 
 async fn test_stream_filter_error_fetch(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -773,7 +764,7 @@ async fn test_stream_filter_error_fetch(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -790,7 +781,10 @@ async fn test_stream_filter_error_fetch(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -909,7 +903,6 @@ async fn test_stream_filter_max_generic() {
 
 /// test filter with max bytes
 async fn test_stream_filter_max(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -919,7 +912,7 @@ async fn test_stream_filter_max(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -936,7 +929,10 @@ async fn test_stream_filter_max(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     // write 2 batches each with 10 records
     //debug!("records: {:#?}", records);
@@ -1056,7 +1052,6 @@ async fn test_stream_fetch_map_adhoc() {
 }
 
 async fn test_stream_fetch_map(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -1067,7 +1062,7 @@ async fn test_stream_fetch_map(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -1083,7 +1078,10 @@ async fn test_stream_fetch_map(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -1227,7 +1225,6 @@ async fn test_stream_fetch_map_error_generic() {
 }
 
 async fn test_stream_fetch_map_error(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -1238,7 +1235,7 @@ async fn test_stream_fetch_map_error(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -1254,7 +1251,10 @@ async fn test_stream_fetch_map_error(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -1379,7 +1379,6 @@ async fn test_stream_aggregate_fetch_single_batch_generic() {
 }
 
 async fn test_stream_aggregate_fetch_single_batch(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -1389,7 +1388,7 @@ async fn test_stream_aggregate_fetch_single_batch(
     let port = portpicker::pick_unused_port().expect("No free ports left");
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -1403,7 +1402,10 @@ async fn test_stream_aggregate_fetch_single_batch(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -1537,7 +1539,6 @@ async fn test_stream_aggregate_fetch_multiple_batch_generic() {
 }
 
 async fn test_stream_aggregate_fetch_multiple_batch(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -1547,7 +1548,7 @@ async fn test_stream_aggregate_fetch_multiple_batch(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -1561,7 +1562,10 @@ async fn test_stream_aggregate_fetch_multiple_batch(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     // Aggregate 6 records in 2 batches
     // First batch:
@@ -1672,7 +1676,6 @@ async fn test_stream_fetch_and_new_request_adhoc() {
 }
 
 async fn test_stream_fetch_and_new_request(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -1682,7 +1685,7 @@ async fn test_stream_fetch_and_new_request(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -1697,7 +1700,10 @@ async fn test_stream_fetch_and_new_request(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -1733,7 +1739,7 @@ async fn test_stream_fetch_invalid_wasm_module_legacy() {
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
     let wasm = Vec::from("Hello, world, I'm not a valid WASM module!");
     let wasm_payload = LegacySmartModulePayload {
         wasm: SmartModuleWasmCompressed::Raw(wasm),
@@ -1741,7 +1747,7 @@ async fn test_stream_fetch_invalid_wasm_module_legacy() {
         ..Default::default()
     };
 
-    test_stream_fetch_invalid_wasm_module(ctx, test_path, Some(wasm_payload), None).await
+    test_stream_fetch_invalid_wasm_module(test_path, Some(wasm_payload), None).await
 }
 
 #[fluvio_future::test(ignore)]
@@ -1750,7 +1756,8 @@ async fn test_stream_fetch_invalid_wasm_module_adhoc() {
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
+
     let wasm = zip(Vec::from("Hello, world, I'm not a valid WASM module!"));
     let smartmodule = SmartModuleInvocation {
         wasm: SmartModuleInvocationWasm::AdHoc(wasm),
@@ -1758,7 +1765,7 @@ async fn test_stream_fetch_invalid_wasm_module_adhoc() {
         ..Default::default()
     };
 
-    test_stream_fetch_invalid_wasm_module(ctx, test_path, None, Some(smartmodule)).await
+    test_stream_fetch_invalid_wasm_module(test_path, None, Some(smartmodule)).await
 }
 
 #[fluvio_future::test(ignore)]
@@ -1767,7 +1774,7 @@ async fn test_stream_fetch_invalid_wasm_module_predefined() {
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
 
     let wasm = zip(Vec::from("Hello, world, I'm not a valid WASM module!"));
     smartmodule_localstore().insert(SmartModule {
@@ -1787,11 +1794,10 @@ async fn test_stream_fetch_invalid_wasm_module_predefined() {
         ..Default::default()
     };
 
-    test_stream_fetch_invalid_wasm_module(ctx, test_path, None, Some(smartmodule)).await
+    test_stream_fetch_invalid_wasm_module(test_path, None, Some(smartmodule)).await
 }
 
 async fn test_stream_fetch_invalid_wasm_module(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -1801,7 +1807,7 @@ async fn test_stream_fetch_invalid_wasm_module(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -1816,7 +1822,10 @@ async fn test_stream_fetch_invalid_wasm_module(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -1900,7 +1909,6 @@ async fn test_stream_fetch_array_map_generic() {
 }
 
 async fn test_stream_fetch_array_map(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -1910,7 +1918,7 @@ async fn test_stream_fetch_array_map(
     let port = portpicker::pick_unused_port().expect("No free ports left");
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -1925,7 +1933,10 @@ async fn test_stream_fetch_array_map(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     // Input: One JSON record with 10 ints: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
     let mut records = BatchProducer::builder()
@@ -2033,7 +2044,6 @@ async fn test_stream_fetch_filter_map_generic() {
 }
 
 async fn test_stream_fetch_filter_map(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -2043,7 +2053,7 @@ async fn test_stream_fetch_filter_map(
     let port = portpicker::pick_unused_port().expect("No free ports left");
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -2058,7 +2068,10 @@ async fn test_stream_fetch_filter_map(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     // Input: the following records:
     //
@@ -2168,7 +2181,6 @@ async fn test_stream_fetch_filter_with_params_generic() {
 }
 
 async fn test_stream_fetch_filter_with_params(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -2179,7 +2191,7 @@ async fn test_stream_fetch_filter_with_params(
     let port = portpicker::pick_unused_port().expect("No free ports left");
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -2195,7 +2207,10 @@ async fn test_stream_fetch_filter_with_params(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx()
+        .leaders_state()
+        .insert(test_id, replica.clone())
+        .await;
 
     let mut params = BTreeMap::new();
     params.insert("key".to_string(), "b".to_string());
@@ -2326,7 +2341,8 @@ async fn test_stream_fetch_invalid_smartmodule_legacy() {
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
+
     let wasm = include_bytes!("test_data/filter_missing_attribute.wasm").to_vec();
     let wasm_payload = LegacySmartModulePayload {
         wasm: SmartModuleWasmCompressed::Raw(wasm),
@@ -2334,7 +2350,7 @@ async fn test_stream_fetch_invalid_smartmodule_legacy() {
         ..Default::default()
     };
 
-    test_stream_fetch_invalid_smartmodule(ctx, test_path, Some(wasm_payload), None).await
+    test_stream_fetch_invalid_smartmodule(test_path, Some(wasm_payload), None).await
 }
 
 #[fluvio_future::test(ignore)]
@@ -2343,7 +2359,8 @@ async fn test_stream_fetch_invalid_smartmodule_adhoc() {
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
+
     let wasm = zip(include_bytes!("test_data/filter_missing_attribute.wasm").to_vec());
     let smartmodule = SmartModuleInvocation {
         wasm: SmartModuleInvocationWasm::AdHoc(wasm),
@@ -2351,7 +2368,7 @@ async fn test_stream_fetch_invalid_smartmodule_adhoc() {
         ..Default::default()
     };
 
-    test_stream_fetch_invalid_smartmodule(ctx, test_path, None, Some(smartmodule)).await
+    test_stream_fetch_invalid_smartmodule(test_path, None, Some(smartmodule)).await
 }
 
 #[fluvio_future::test(ignore)]
@@ -2360,7 +2377,7 @@ async fn test_stream_fetch_invalid_smartmodule_predefined() {
     let mut spu_config = SpuConfig::default();
     spu_config.log.base_dir = test_path.clone();
 
-    let ctx = GlobalContext::new_shared_context(spu_config);
+    initialize(spu_config);
 
     let wasm = zip(include_bytes!("test_data/filter_missing_attribute.wasm").to_vec());
     smartmodule_localstore().insert(SmartModule {
@@ -2380,11 +2397,10 @@ async fn test_stream_fetch_invalid_smartmodule_predefined() {
         ..Default::default()
     };
 
-    test_stream_fetch_invalid_smartmodule(ctx, test_path, None, Some(smartmodule)).await
+    test_stream_fetch_invalid_smartmodule(test_path, None, Some(smartmodule)).await
 }
 
 async fn test_stream_fetch_invalid_smartmodule(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -2394,7 +2410,7 @@ async fn test_stream_fetch_invalid_smartmodule(
     let port = portpicker::pick_unused_port().expect("No free ports left");
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -2409,7 +2425,7 @@ async fn test_stream_fetch_invalid_smartmodule(
     let replica = LeaderReplicaState::create(test, config(), status_update_owned())
         .await
         .expect("replica");
-    ctx.leaders_state().insert(test_id, replica.clone()).await;
+    default_replica_ctx().leaders_state().insert(test_id, replica.clone()).await;
 
     let stream_request = DefaultStreamFetchRequest {
         topic: topic.to_owned(),
@@ -2446,7 +2462,6 @@ async fn test_stream_fetch_invalid_smartmodule(
 }
 
 async fn test_stream_fetch_join(
-    ctx: Arc<GlobalContext<FileReplica>>,
     test_path: PathBuf,
     wasm_payload: Option<LegacySmartModulePayload>,
     smartmodule: Option<SmartModuleInvocation>,
@@ -2462,7 +2477,7 @@ async fn test_stream_fetch_join(
 
     let addr = format!("127.0.0.1:{}", port);
 
-    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+    let server_end_event = create_public_server(addr.to_owned()).run();
 
     // wait for stream controller async to start
     sleep(Duration::from_millis(100)).await;
@@ -2481,7 +2496,7 @@ async fn test_stream_fetch_join(
         LeaderReplicaState::create(test_left.clone(), config(), status_update_owned())
             .await
             .expect("replica");
-    ctx.leaders_state()
+    default_replica_ctx().leaders_state()
         .insert(test_id_left, replica_left.clone())
         .await;
 
@@ -2494,7 +2509,7 @@ async fn test_stream_fetch_join(
             .await
             .expect("replica");
     replica_localstore().insert(test_right);
-    ctx.leaders_state()
+    default_replica_ctx().leaders_state()
         .insert(test_id_right, replica_right.clone())
         .await;
